pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eiml"
version = "0.1.0"
description = "EIML descriptors (SOAP/EIML) + SGPR training utilities"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "pyyaml",
  "ase",
  "scikit-learn",
]

[project.scripts]
eiml = "eiml.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
